+++ domain/navigation/src/main/java/com/ku_stacks/ku_ring/navigation/KuringRoute.kt
@@ -1,0 +1,6 @@
+package com.ku_stacks.ku_ring.navigation
+
+interface KuringRoute {
+    val route: String
+        get() = this::class.java.canonicalName.toString()
+}
