+++ feature/main/src/main/java/com/ku_stacks/ku_ring/main/MainScreenNavigationBar.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.unit.sp
 import com.ku_stacks.ku_ring.designsystem.components.LightAndDarkPreview
 import com.ku_stacks.ku_ring.designsystem.kuringtheme.KuringTheme
 import com.ku_stacks.ku_ring.designsystem.kuringtheme.values.Pretendard
+import com.ku_stacks.ku_ring.navigation.MainScreenRoute
 
 @Composable
 internal fun MainScreenNavigationBar(
